+++ cpp/cpp_odbc/Test/cpp_odbc_test/mock_statement.h
@@ -7,8 +7,8 @@ namespace cpp_odbc_test {
 
 class mock_statement : public cpp_odbc::statement {
 public:
-	MOCK_CONST_METHOD1( do_get_integer_attribute, long(SQLINTEGER));
-	MOCK_CONST_METHOD2( do_set_attribute, void(SQLINTEGER, long));
+	MOCK_CONST_METHOD1( do_get_integer_attribute, intptr_t(SQLINTEGER));
+	MOCK_CONST_METHOD2( do_set_attribute, void(SQLINTEGER, intptr_t));
 	MOCK_CONST_METHOD2( do_set_attribute, void(SQLINTEGER, SQLULEN *));
 	MOCK_CONST_METHOD1( do_execute, void(std::string const &));
 	MOCK_CONST_METHOD1( do_prepare, void(std::string const &));
@@ -22,7 +22,7 @@ public:
 	MOCK_CONST_METHOD0( do_unbind_all_columns, void());
 	MOCK_CONST_METHOD0( do_fetch_next, bool());
 	MOCK_CONST_METHOD0( do_close_cursor, void());
-	MOCK_CONST_METHOD2( do_get_integer_column_attribute, long(SQLUSMALLINT, SQLUSMALLINT));
+	MOCK_CONST_METHOD2( do_get_integer_column_attribute, intptr_t(SQLUSMALLINT, SQLUSMALLINT));
 	MOCK_CONST_METHOD2( do_get_string_column_attribute, std::string(SQLUSMALLINT, SQLUSMALLINT));
 	MOCK_CONST_METHOD0( do_row_count, SQLLEN());
 	MOCK_CONST_METHOD1( do_describe_column, cpp_odbc::column_description(SQLUSMALLINT));
